Extract ghosting utilities from NonConformalManager
The ghosting add/remove logic does not depend on NonConformalManager
class and could be utilized in other places where ghosting changes as a
function of timestep (e.g., overset, actuator line modeling, etc.). This
commit extracts the common pieces and consolidates those functions as
part of StkHelpers so that they can be used across different parts of
Nalu-Wind.
