+++ include/NonConformalManager.h
@@ -64,11 +64,6 @@ class NonConformalManager {
   stk::mesh::EntityProcVec elemsToGhost_;
   std::vector<NonConformalInfo *> nonConformalInfoVec_;
 
-  static void compute_precise_ghosting_lists(const stk::mesh::BulkData& bulk,
-                                             stk::mesh::EntityProcVec& elemsToGhost,
-                                             stk::mesh::EntityProcVec& curSendGhosts,
-                                             std::vector<stk::mesh::EntityKey>& recvGhostsToRemove);
-
   std::vector<int> ghostCommProcs_;
 
   private:
