+++ src/CMakeLists.txt
@@ -219,6 +219,7 @@ add_subdirectory(overset)
 add_subdirectory(pmr)
 add_subdirectory(property_evaluator)
 add_subdirectory(tabular_props)
+add_subdirectory(utils)
 add_subdirectory(user_functions)
 add_subdirectory(xfer)
 add_subdirectory(wind_energy)
