+++ src/utils/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_sources(GlobalSourceList
+  StkHelpers.C
+  )
